still have a lot of bugs. but this is the last commit.

+++ Classes/GameObject/BulletObject.h
@@ -23,7 +23,7 @@ public:
 	Vec2 vector;
 	void bulletFire(Vec2 location);
 protected:
-	int _Dmg;
+	float _Dmg;
 private:
 	void poolAndRemoveBullet();
 private:
